requirements.txt
discord.py>=2.5.2
# git+https://github.com/Rapptz/discord.py/
motor>=3.7.0
pymongo>=4.12.0
pandas>=2.2.3
apscheduler>=3.11.0
matplotlib>=3.10.1
requests>=2.32.3
pytest>=8.3.5
